Demo/requirements.txt
langchain==0.0.180
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.20.0
faiss-cpu==1.7.4
sentence_transformers==2.2.2
chromadb==0.3.25
tiktoken==0.4.0
streamlit-extras
openai==0.27.7
pypdf==3.9.0